+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Go Buildpack Changelog
 
 ## Unreleased
+* Drop Heroku-16 from CI test matrix
 
 ## v153 (2021-03-11)
 * Add go1.16.1, use for go1.16